\name{ADMUR} 
\alias{ADMUR} 
\title{ADMUR R package} 
\docType{package}
\description{
ADMUR: Ancient Demographic Modelling Using Radiocarbon

Bayesia tools to directly model and compare the underlying dynamics of prehistoric population change using radiocarbon date datasets.
	}
\details{
	
Package to model population dynamics from anthropogenic datasets of dates such as radiocarbon dates. 

Models the underlying population dynamics using common parameteric models, including uniform, gaussian, exponential, logistic, sinusoidal, cauchy, power, etc.

Models the underlying population dynamics using a Continuous Piecewise Linear (CPL) model framework, including estimating the dates and magnitude of demographic events (hinge points).

Models the underlying population dynamics using an independent timeseries. 

Generates Summed Probability Distribution (SPD) and tests a null hypothesis using simulation.

Tools to estimate Maximum likelihood; posterior distributions; Goodness of Fit tests; model comparison.

Tools to model archaeological phases, and estimate date ranges.
	}
\references{
	'Directly modelling population dynamics in the South American Arid Diagonal using 14C dates' by Adrian Timpson, Ramiro Barberena, Mark G. Thomas, César Méndez and Katie Manning, published in Philosophical Transactions of the Royal Society B, 2020. https://doi.org/10.1098/rstb.2019.0723
	
	'ADMUR: Ancient Demographic Modelling Using Radiocarbon. Adrian Timpson. University College London. 
	Research Department of Genetics, Environment and Evolution (GEE), Darwin Building, Gower Street, London, WC1E 6BT. 2020 
	https://github.com/UCL/ADMUR
	}
