\name{bluhm2421} 
\docType{data}
\alias{bluhm2421} 
\title{Radiocarbon dataset from Bluhm and Surovell 2018} 
\description{
	The 14C dataset used by Bluhm and Surovell in 'Validation of a global model of taphonomic bias using geologic radiocarbon', Quaternary Research 2018.
	Data provided from \href{https://www.cambridge.org/core/journals/quaternary-research/article/abs/validation-of-a-global-model-of-taphonomic-bias-using-geologic-radiocarbon-ages/42E38E070B2B813D55BCEE8D9DC0DF93#supplementary-materials}{Supplementary materials} originally comprised 2457 rows, but 36 rows (dates( have no mean or SD (only a minimum age) and have been excluded from this data frame. Furthermore, 25 rows (dates) have to SDs (+ve and -ve), which have been changed to a single SD for this dataset by taking the mean average.
	}
\usage{
	bluhm2421
	}
\format{
	A data frame comprising 2421 rows and 4 columns. 
	}

