\name{bryson1848} 
\docType{data}
\alias{bryson1848} 
\title{Radiocarbon dataset from Bryson et al. 2006} 
\description{
	The 14C dataset used by Surovell et al. in 'Correcting temporal frequency distributions for taphonomic bias', Journal of Archaeological Science 2009.
	Data was originally published by Bryson et al. in 'A calibrated radiocarbon database of late Quaternary volcanic eruptions', Earth Discussions, 2006.
	Bryson et al. \href{http://www.electronic-earth-discuss.net/1/123/2006/eed-1-123-2006-supplement.zip}{source data} contains two data tables. The second comprises 173 rows of dates missing their standard deviations. Instead this data frame corresponds to Brysons first table comprising 1848 rows. However, these  are often an amalgamation of several individual radiocarbon dates which cannot be reverse-engineered to obtain the individual dates. Therefore we assume each row represents a unique phase.
	}
\usage{
	bryson1848
	}
\format{
	A data frame comprising 1848 rows and 8 columns. 
	}

