\name{data4} 
\docType{data}
\alias{data4} 
\title{Toy radiocarbon dataset} 
\description{
	Data frame comprising a randomly generated 14C dataset. The data generation process is deliberately left to the imagination.
	}
\usage{
	data4
	}
\format{
	A data frame comprising 100 rows and 4 columns. 
	}

