\name{data5} 
\docType{data}
\alias{data5} 
\title{Toy radiocarbon dataset} 
\description{
	Data frame comprising a randomly generated 14C dataset. The true data generation process is deliberately left to the imagination
	}
\usage{
	data5
	}
\format{
	A data frame comprising 350 rows and 4 columns. 
	}

