\name{data6} 
\docType{data}
\alias{data6} 
\title{Toy radiocarbon dataset} 
\description{
	Data frame comprising a randomly generated 14C dataset. The true data generation process is deliberately left to the imagination
	}
\usage{
	data6
	}
\format{
	A data frame comprising 57 rows and 5 columns. 
	}

