\name{dateCalibrator} 
\alias{dateCalibrator} 
\title{Generates an calibrated PD for each date in a dataset} 
\description{Generates a matrix of calibrated date probabilities, each column corresponding to each uncalibrated date}
\usage{dateCalibrator(data, CalArray)}
\arguments{
  \item{data}{
	A dataframe of 14C dates comprising 'age' and 'sd'.
	}
  \item{CalArray}{
	A 2D probability array of the calibration curve generated by \code{\link[=makeCalArray]{makeCalArray}} containing row names and column names. 
	}
}
\value{
	Returns a matrix of probabilities. Each column provides the calibrated probability density of a date. Row names are the calendar years.
	}

\examples{

	# minimum data requirement includes 'mean' and 'sd'
	data <- data.frame(age=c(5800, 5100, 5500, 5320, 5560),sd=c(40, 35, 30, 30, 30))
	calrange <- estimateDataDomain(data, intcal20)
	CalArray <- makeCalArray(intcal20, calrange)
	x <- dateCalibrator(data, CalArray)
	}
