\name{intcal13} 
\docType{data}
\alias{intcal13} 
\title{Northern hemisphere 2013 calibration curve} 
\description{
	Northern hemisphere 2013 calibration curve
	}
\usage{
	intcal13
	}
\format{
	A data frame comprising 5141 rows and 3 columns:
	cal BP, 14C BP, +/- Error.
	}
\source{
	Atmospheric data from Reimer et al (2013). Obtained from the raw intcal13.14c file downloaded from http://radiocarbon.webhost.uits.arizona.edu/node/19
	}
\references{
	Reimer PJ, Bard E, Bayliss A, Beck JW, Blackwell PG, Bronk Ramsey C, Buck CE, Cheng H, Edwards RL, Friedrich M, Grootes PM, Guilderson TP, Haflidason H, Hajdas I, Hatte C, Heaton TJ, Hogg AG, Hughen KA, Kaiser KF, Kromer B, Manning SW, Niu M, Reimer RW, Richards DA, Scott EM, Southon JR, Turney CSM, van der Plicht J. Radiocarbon 55(4). DOI: 10.2458/azu_js_rc.55.16947}
