\name{objectiveFunction} 
\alias{objectiveFunction} 
\title{Objective function to be minimised in a search. Returns the negative log likelihood.} 
\description{Calculates the negative log likelihood of a model given a calibrated date matrix.}
\usage{objectiveFunction(pars, PDarray, type='CPL')}
\arguments{
  \item{pars}{
	A single vector of parameters, each between 0 and 1. Must be an odd length since parameters correspond to: (y,x,...y)
	}
  \item{PDarray}{
	A data frame typically generated by \code{\link[=phaseCalibrator]{phaseCalibrator}}, such that each column represents the PDF of each calibrated date (or phase), and row names are the corresponding years.
	}
  \item{type}{
	Choose from 'CPL', 'exponential', or 'uniform'. 
	}
}
\value{
	Returns a single value, the negative log likelihood.
	}
\details{
	}
\examples{
	# generate a PD array from a handful of dates
data <- subset(arid, site %in% c('Carrizal','Pacopampa'))
	CalArray <- makeCalArray(shcal13, calrange = c(2000,6000))
	PD <- phaseCalibrator(data, CalArray)

	# some random parameters for a 3-CPL model, find negative log likelihood
	pars <- runif(5)
	objectiveFunction(pars, PD)	
}
