\name{plotCal} 
\alias{plotCal} 
\title{Plots a calibrated PD of a single date, or calibrated SPD of multiple dates} 
\description{Generates a basic plot of the Probability Distribution of calibrated date, or Summed Probability Distribution of multiple calibrated dates.}
\usage{plotCal(cal)}
\arguments{
  \item{cal}{
	A data frame comprising columns 'calBP' and 'prob', for example as generated by \code{\link[=summedCalibrator]{summedCalibrator}}
	}
}
\details{
	Presents the calibrated date (or SPD) as a grey polygon.
	}
\examples{
	data <- data.frame(age=c(9144),sd=c(151))
	CalArray <- makeCalArray(intcal13,calrange=c(8000,13000))
	cal <- summedCalibrator(CalArray,data)
	plotCal(cal)
	}
