\name{shcal20} 
\docType{data}
\alias{shcal20} 
\title{Southern hemisphere 2020 calibration curve} 
\description{Southern hemisphere 2020 calibration curve}
\usage{shcal20}
\format{
	A data frame comprising 9501 rows and 3 columns:
	cal BP, 14C BP, +/- Error.
	}
\source{
	Atmospheric data from Hogg et al. (2020). Obtained from the raw shcal20.14c file downloaded from http://intcal.org/curves/shcal20.14c
	}
\references{
	Hogg AG, Heaton TJ, Hua Q, Palmer JG, Turney CSM, Southon J, Bayliss A, Blackwell PG, Boswijk G, Bronk Ramsey C, Pearson C, Petchey F, Reimer P, Reimer R, Wacker L. 2020. SHCal20 Southern Hemisphere calibration, 0-55,000 years cal BP. Radiocarbon 62. doi: 10.1017/RDC.2020.59
	}
	