\name{shcal13} 
\docType{data}
\alias{shcal13} 
\title{Southern hemisphere 2013 calibration curve} 
\description{Southern hemisphere 2013 calibration curve}
\usage{shcal13}
\format{
	A data frame comprising 5141 rows and 3 columns:
	cal BP, 14C BP, +/- Error.
	}
\source{
	Atmospheric data from Hogg et al. (2013). Obtained from the raw shcal13.14c file downloaded from http://radiocarbon.webhost.uits.arizona.edu/node/19
	}
\references{Hogg, A.G., Hua, Q., Blackwell, P.G., Niu, M., Buck, C.E., Guilderson, T.P., Heaton, T.J., Palmer, J.G., Reimer, P.J., Reimer, R.W., 2013. SHCal13 Southern Hemisphere calibration, 0–50,000 years cal BP, Radiocarbon 55, 1889-1903.}

