\name{summedCalibrator} 
\alias{summedCalibrator} 
\title{Generates a summed probability distribution (SPD) of calibrated dates} 
\description{Calibrates 14C dates through the CalArray, then sums.}
\usage{summedCalibrator(CalArray,data)}
\arguments{
  \item{CalArray}{
	A 2D probability array of the calibration curve, with row names and column names. Generated by makeCalArray().
	}
  \item{data}{
	A dataframe of 14C dates. Requires 'age' and 'sd'.
	}
}
\details{
	Uses CalArray once to simultaneously calibrate and sum all 14C dates in data. 
	The result is equivalent to calibrating each date, then summing.
	Note: density values are un-normalised. I.e., the total area of the SPD will equal the number of dates calibrated.
	}
\examples{
	# SPD of three 14C dates
	CalArray <- makeCalArray(intcal13, calrange=c(9000,11000) )
	data <- data.frame(age = c(8560,8900,9150), sd = c(50,50,50) )
	SPD <- summed.calibrator(CalArray, data)
	plot(SPD, type='l')

	# note under the default 5 cal yr resolution the total probability density = 3 / 5
	sum(SPD$prob)

	# At 1 cal yr resolution the total probability density = 3
	CalArray <- makeCalArray(intcal13, calrange=c(9000,11000), inc=1 )
	data <- data.frame(age = c(8560,8900,9150), sd = c(50,50,50) )
	SPD <- summed.calibrator(CalArray, data )
	plot(SPD, type='l')
	sum(SPD$prob)
	}
