\name{toy} 
\docType{data}
\alias{toy} 
\title{Toy population model} 
\description{
	Data frame of a toy population model. Provides a discretised PDF across the time range of 7.5kyr to 5.5kyr BP. The model PDF outside this range is zero. Comprises two columns: year, pdf.
	}
\usage{
	toy
	}
\format{
	A data frame.
	}
\references{
	Timpson, A., Barberena, R., Thomas, M.G., Méndez, C., Manning, K. 2020 Directly modelling population dynamics in the South American Arid Diagonal using 14C dates. Philosophical Transactions of the Royal Society B.
	}