\name{ADMUR} 
\alias{ADMUR} 
\title{ADMUR R package} 
\docType{package}
\description{
ADMUR: Ancient Demographic Modelling Using Radiocarbon

Tools to directly model and compare the underlying dynamics of prehistoric population change using radiocarbon date datasets.
	}
\details{
	Package to model population dynamics from anthropogenic datasets of dates such as radiocarbon dates. Provides tools to perform and compare three modelling approaches:

1. Directly modelling the underlying population dynamics using common parameteric models, including uniform, gaussian, exponential, logistic, sinusoidal, cauchy, power, etc.

2. Directly modelling the underlying population dynamics using a Continuous Piecewise Linear (CPL) model framework, including estimating the dates and magnitude of demographic events (hinge points).

3. Directly modelling the underlying population dynamics using an independent timeseries. 

4. Summed Probability Distribution (SPD) generation, and simulation testing. Permits the rejection of a null hypothesis.

Tools to estimate Maximum likelihood; Goodness of Fit tests; model comparison.
	}
\references{
	'Directly modelling population dynamics in the South American Arid Diagonal using 14C dates' by Adrian Timpson, Ramiro Barberena, Mark G. Thomas, César Méndez and Katie Manning, published in Philosophical Transactions of the Royal Society B, 2020. https://royalsocietypublishing.org/doi/10.1098/rstb.2019.0723
	
	'ADMUR: Ancient Demographic Modelling Using Radiocarbon. Adrian Timpson. University College London. 
	Research Department of Genetics, Environment and Evolution (GEE), Darwin Building, Gower Street, London, WC1E 6BT. 2020 
	https://github.com/UCL/ADMUR
	}
