\name{data3} 
\docType{data}
\alias{data3} 
\title{Toy radiocarbon dataset} 
\description{
	Data frame comprising a randomly generated 14C dataset. The true data generation process is deliberately left to the imagination
	}
\usage{
	data3
	}
\format{
	A data frame comprising 100 rows and 4 columns. 
	}

