\name{summedCalibrator} 
\alias{summedCalibrator} 
\title{Generates a summed probability distribution (SPD) of calibrated dates} 
\description{Generates a Summed Probability Distribution (SPD). Handles both C14 and other date types e.g., thermoluminescence.}
\usage{summedCalibrator(data, CalArray, normalise = TRUE)}
\arguments{
  \item{data}{
	A dataframe of 14C dates. Requires 'age' and 'sd', and at least one of 'site' and 'phase'. Optional 'datingType' to include '14C' and anything else.
	}
  \item{CalArray}{
	A 2D probability array of the calibration curve generated by \code{\link[=makeCalArray]{makeCalArray}} containing row names and column names. 
	}
  \item{normalise}{
	Default TRUE normalises probability distribution to give a total area equal to 1. Otherwise if FALSE total PD area equal to the number of dates. 
	}
}
\value{
	Returns a single-column data frame of SPD probabilities. Row names are the calendar years.
	}
\details{
Uses CalArray once to simultaneously calibrate and sum all 14C dates in data. The result is equivalent to calibrating each date, then summing. Note, if normalise = TRUE the SPD total area equals 1, provided the date range of the CalArray encompasses the dates provided. If the date range used to generate CalArray does not encompass the entire dataset, some dates will have PD outside the date range, giving a SPD area < 1. This avoids deleterious edge effects. Similarly if normalise = FALSE the SPD total may be < the number of dates.

Optionally 'datingType' can be provided in the data. Only '14C' will be calibrated in the usual way, anything else is assumed to be provided in calendar time. If 'datingType' is not provided, all dates are assumed to be 14C.

	}
\examples{
	# SPD of three 14C dates
	CalArray <- makeCalArray(intcal20, calrange=c(9000,11000) )
	data <- data.frame(
		age = c(8350,8900,9350), 
		sd = rep(50,3),
		datingType = rep('14C',3)
		)
	SPD <- summedCalibrator(data, CalArray, normalise = FALSE)
	plotPD(SPD)

	# At 5 cal yr resolution the SPD area = 3
	sum(SPD) * 5

	# At 1 cal yr resolution the SPD area also = 3
	CalArray <- makeCalArray(intcal20, calrange=c(9000,11000), inc=1 )
	SPD <- summedCalibrator(data, CalArray, normalise = FALSE)
	plotCal(SPD)
	sum(SPD) * 1

	# and under the default normalisation the SPD area = 1
	SPD <- summedCalibrator(data, CalArray, normalise = TRUE)
	plotPD(SPD)
	sum(SPD) * 1
	
	# whilst a date outside the date range results in the SPD area < 1
	data <- data.frame(
		age = c(8350,8900,11000), 
		sd = rep(50,3),
		datingType = c('14C','14C','TL')
		)
	SPD <- summedCalibrator(data, CalArray, normalise = TRUE)
	plotPD(SPD)
	sum(SPD) * 1
	}
