\name{summedCalibrator} 
\alias{summedCalibrator} 
\title{Generates a summed probability distribution (SPD) of calibrated dates} 
\description{Generates a Summed Probability Distribution (SPD). Handles both C14 and other date types e.g., thermoluminescence.}
\usage{summedCalibrator(data, CalArray, normalise = 'standard')}
\arguments{
  \item{data}{
	A dataframe of 14C dates. Requires 'age' and 'sd', and at least one of 'site' and 'phase'. Optional 'datingType' to include '14C' and anything else.
	}
  \item{CalArray}{
	A 2D probability array of the calibration curve generated by \code{\link[=makeCalArray]{makeCalArray}} containing row names and column names. 
	}
  \item{normalise}{
	One of 'none', 'standard', or 'full'. 
	}
}
\value{
	Returns a single-column data frame of SPD probabilities. Row names are the calendar years.
	}
\details{
Uses CalArray once to simultaneously calibrate and sum all 14C dates in data. The result is equivalent to calibrating each date, then summing. 

Optionally 'datingType' can be provided in the data. Only '14C' will be calibrated in the usual way, anything else is assumed to be provided in calendar time. If 'datingType' is not provided, all dates are assumed to be 14C.

If normalise is 'none', the output is a PD with an area equal to the total number of samples. This option is rarely required, but can be useful for example when plotting several SPDs and would also like to illustrate that one dataset is larger than another. However, if the date range used to generate CalArray does not encompass the entire dataset, some dates will have some probability mass outside the date range. Therefore the total probability can be less than the sample size to avoid deleterious edge effects. 

If normalise is 'standard', the output is adjusted by the number of samples, such that the total area equals 1. However, as with the 'none' option, if the date range used to generate CalArray does not encompass the entire dataset, some dates will have some probability mass outside the date range. Therefore the total probability can be less than the sample size to avoid deleterious edge effects.

If normalise is 'full', the output PD always has an area equal to 1, even if there are dates with a substantial probability mass outside the date range. This option should only be used when all dates are of equal importance, and the resulting SPD is the final product. For example when generating simulated SPDs.

	}
\examples{
	# SPD of three 14C dates
	CalArray <- makeCalArray(intcal20, calrange=c(9000,10650), inc=1 )
	data <- data.frame(
		age = c(8350,8900,9350), 
		sd = rep(50,3),
		datingType = rep('14C',3)
		)


	# with the default normalisation the SPD area is a little under 1
	# since the date range is too narrow
	SPD <- summedCalibrator(data, CalArray)
	plotPD(SPD)
	sum(SPD) 

	# without normalisation the total area is a little under 3
	SPD <- summedCalibrator(data, CalArray, normalise='none')
	plotPD(SPD)
	sum(SPD)

	# with full normalisation the total area is exactly 1
	SPD <- summedCalibrator(data, CalArray, normalise='full')
	plotPD(SPD)
	sum(SPD)
	}
